Keep graph explorer CI checks green
